factored out eslint result filter to its own module
