+++ tasks/eslint/details.js
@@ -2,25 +2,8 @@
 
 var gulp = require('gulp');
 var eslint = require('gulp-eslint');
-var getArgs = require('minimist');
-var TransformStream = require('stream').Transform;
 
-var messageFilter = require('./message-filter');
-
-var argv = getArgs(process.argv.slice(2));
-
-function filterResults () {
-    var stream = new TransformStream({ objectMode: true });
-
-    stream._transform = function (file, enc, cb) {
-        if (file.eslint && (argv.rule || argv.keyword)) {
-            file.eslint = messageFilter(file.eslint, argv.rule, argv.keyword);
-        }
-        cb(null, file);
-    };
-
-    return stream;
-}
+var filterResults = require('./result-filter');
 
 module.exports = function taskFactory (src, bail) {
